Add CSS update to label, meta html tag-viewport, JS working again after breaking

+++ main.js
@@ -8,11 +8,13 @@ var favoriteBtn = document.querySelector(".favorite-btn");
 var addBtn = document.querySelector(".add-btn");
 var photoArea = document.querySelector(".photo-card-area");
 var photoArray = [];
+var favoriteImage = [];
+var reader = new FileReader();
 
 //Event Listeners
-searchBtn.addEventListener('click', searchPhotos);
+searchInput.addEventListener('keyup', searchPhotos);
 addBtn.addEventListener('click', createCard);
-// favoriteBtn.addEventListener();
+// favoriteBtn.addEventListener('click', changeHeart);
 
 window.onload = function() {
   var keys = Object.keys(localStorage);
@@ -24,8 +26,8 @@ window.onload = function() {
   })
 }
 
-function createCard (event) {
-  event.preventDefault();
+function createCard (e) {
+  e.preventDefault();
   var titleInput = document.querySelector('.title-input');
   var captionInput = document.querySelector('.caption-input');
   const newPhoto = new Photo(Date.now(), titleInput.value, captionInput.value);
@@ -39,39 +41,50 @@ function createCard (event) {
   }
 }
 
+// function changeHeart() {
+//   if (this.favorite === false) {
+//     console.log(15);
+//     element.src = "images/favorite-active.svg"
+//   }
+// }
+
+
 function appendCard(photo) {
   var card = 
     `<article class="card" data-id="${photo.id}">
       <header class="card-head">
-        <h4 class="card-head-text">${photo.title}</h4>
+        <h4 class="card-head-text" contenteditable="true">${photo.title}</h4>
       </header>
-      <section class="card-pic"></section>
+      <section class="card-pic" background-image=url("${photo.file}")></section>
       <section class="card-body">
-        <p class="card-body-text">${photo.caption}</p>
+        <p class="card-body-text" contenteditable="true">${photo.caption}</p>
       </section>
       <footer class="card-foot">
         <img src="images/delete.svg" onclick="deleteCard(${photo.id})">
-        <img src="images/favorite.svg">
+        <img src="images/favorite.svg" onclick="">
       </footer>
     </article>`
     photoArea.innerHTML = card + photoArea.innerHTML;
 }
 
 function deleteCard (id) {
-  let element = document.querySelector(`[data-id="${id}"]`);
-  element.remove();
-  let deleteIdea = photoArray.find(function(photo) {
-    return id === photo.id;
-  });
-  deleteIdea.deleteFromStorage();
-  let deleteIndex = photoArray.findIndex(function(photo) {
-    return id === photo.id;
-  });
-  photoArray.splice(deleteIndex, 1)
+  var result = confirm("Are you sure you want to delete this photograph?");
+  if (result) {
+    let element = document.querySelector(`[data-id="${id}"]`);
+    element.remove();
+    let deleteIdea = photoArray.find(function(photo) {
+      return id === photo.id;
+    });
+    deleteIdea.deleteFromStorage();
+    let deleteIndex = photoArray.findIndex(function(photo) {
+      return id === photo.id;
+    });
+    photoArray.splice(deleteIndex, 1)
+  }
 }
 
-function searchPhotos (event) {
-  event.preventDefault();
+function searchPhotos (e) {
+  e.preventDefault();
   var searchWord = searchInput.value.toUpperCase();
   var filteredPhotos = photoArray.filter(function(obj) {
     var titleText = obj.title.toUpperCase();
@@ -84,5 +97,36 @@ function searchPhotos (event) {
   })
 }
 
+//Travis code
+// var create = document.querySelector('.add-btn');
+// var input = document.querySelector('.choose-file-btn');
+// var photoSection = document.querySelector('.card-pic');
+// var imagesArr = JSON.parse(localStorage.getItem('photos')) || [];
+
+
+// window.addEventListener('load', appendPhotos);
+// create.addEventListener('click', createElement);
+
+// function appendPhotos() {
+//   imagesArr.forEach(function (photo) {
+//     photoSection.innerHTML += `<img src=${photo.file} />`
+//   })
+// }
+
+// function createElement(e) {
+//   console.log(input)
+//   if (input.files[0]) {
+//     reader.readAsDataURL(input.files[0]); 
+//     reader.onload = createCard;
+//   }
+// }
+
+// function addPhoto(e) {
+//   console.log(e.target.result);
+//   var newPhoto = new Photo(Date.now(), e.target.result);
+//   photoSection.innerHTML += `<img src=${e.target.result} />`;
+//   imagesArr.push(newPhoto)
+//   newPhoto.saveToStorage(imagesArr)
+// }
